requirements.txt
albumentations==0.5.2
braceexpand==0.1.7
bytecode==0.13.0.dev0
diffusers==0.27.2
easydict==1.9
Flask==3.0.3
gradio==4.31.0
groundingdino==0.1.0
huggingface_hub==0.21.0
imageio==2.34.1
imageio_ffmpeg==0.4.9
imgaug==0.4.0
joblib==1.4.2
jpeg4py==0.1.4
kornia==0.5.0
lightning==2.2.4
lmdb==1.4.1
MarkupSafe==2.1.5
matplotlib==3.8.4
memory_profiler==0.61.0
numpy==1.26.4
omegaconf==2.1.2
onnxruntime==1.17.3
opencv_python==4.9.0.80
pandas==2.2.2
Pillow==9.3.0
Pillow==10.3.0
pycocotools==2.0.7
pytorch_lightning==1.2.9
PyYAML==6.0.1
PyYAML==6.0.1
scikit_learn==1.4.2
scipy==1.13.0
setuptools==59.6.0
six==1.16.0
skimage==0.0
sync_batchnorm==0.0.1
tensorboardX==2.6.2.2
tensorflow==2.16.1
thop==0.1.1.post2209072238
timm==0.9.16
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
webdataset==0.2.86
